+++ client/src/index.js
@@ -16,11 +16,11 @@ dotenv.config();
 //console.log("******",process.env.REACT_APP_API, process.env.NODE_ENV, "******")
 
 // usando el back deploy ------------------------------------
-// axios.defaults.baseURL = process.env.REACT_APP_API;
+axios.defaults.baseURL = process.env.REACT_APP_API;
 //-----------------------------------------------------------
 
 // usando el back LOCAL -------------------------------------
-axios.defaults.baseURL = "http://localhost:3001";
+// axios.defaults.baseURL = "http://localhost:3001";
 //-----------------------------------------------------------
 
 ReactDOM.render(
